third_party/blink/web_tests/platform/generic/fast/multicol/inline-getclientrects-expected.txt

PASS rects.length is 2
PASS rects[1].left is > rects[0].left
PASS rects[0].top is > rects[1].top
PASS rects[0].right - rects[0].left is > 0
PASS rects[1].right - rects[1].left is rects[0].right - rects[0].left
PASS rects[0].bottom - rects[0].top is > 0
PASS rects[1].bottom - rects[1].top is rects[0].bottom - rects[0].top
PASS successfullyParsed is true

TEST COMPLETE






XXXXXXXXXXXXXXXXXXXXXXXXXX XXXXXXXXXXXXXXXXXXXXXXXXXX
